frontend/requirements.txt
streamlit
ultralytics
requests
python-dotenv
Pillow